Add LWAPI static assert size and OpenLW std macros
OPENLW_STD_NAMESPACE expands to ::std:: normally, and can be defined to expand to something else. Useful for using vs2010 stl types from modern vs, for example.

LWAPI_STATIC_ASSERT_SIZE does what it sounds like; ensures that the given type is the given size at compile-time. It's not used in OpenLW (we don't want to require the sizes to be exact in the recomp), but will be used in LWAPI.